+++ packages/template/src/type.ts
@@ -1,17 +1,31 @@
+export type Param = {
+  name: string
+  label?: string
+  type: 'select' | 'input' | 'boolean'
+  options: string[]
+}
+export type Select = [string, string[]]
+export type Boolean = [string, 'boolean']
+
 export interface TemplateOption {
   name: string
   root: string
-  params: string[]
+  params: (string | Select | Boolean)[]
   templates: Template[]
 }
 
-export interface Template {
-  name:  string;
-  type:  string;
-  files: TemplateFile[];
+export type ReplaceOption = {
+  target: string
+  template?: string
+  content?: string
 }
 
-export interface TemplateFile {
-  name:     string;
-  template: string;
+export interface Template {
+  name: string;
+  type: 'dir' | 'file' | 'replace';
+  replaceFile?: string;
+  replaceOptions?: ReplaceOption[];
+  root: string;
+  template?: string;
+  files?: Template[];
 }