+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @zdcode/cli
 
+## 2.2.1
+
+### Patch Changes
+
+- fix: 加上冒号匹配失败问题
+- Updated dependencies
+  - @zdcode/template@2.2.1
+
+## 2.2.0
+
+### Minor Changes
+
+- 支持创建多个目录文件、在文件中插入
+
+### Patch Changes
+
+- Updated dependencies
+  - @zdcode/template@2.2.0
+
+## 2.1.1
+
+### Patch Changes
+
+- 修复正则 bug
+- Updated dependencies
+  - @zdcode/template@2.1.1
+
+## 2.1.0
+
+### Minor Changes
+
+- 支持驼峰转链式
+
 ## 2.0.0
 
 ### Major Changes