+++ packages/template/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @zdcode/template
 
+## 2.2.1
+
+### Patch Changes
+
+- fix: 加上冒号匹配失败问题
+
+## 2.2.0
+
+### Minor Changes
+
+- 支持创建多个目录文件、在文件中插入
+
+## 2.1.1
+
+### Patch Changes
+
+- 修复正则 bug
+
+## 2.1.0
+
+### Minor Changes
+
+- 支持驼峰转链式
+
+### Patch Changes
+
+- Updated dependencies
+  - @zdcode/utils@2.1.0
+
 ## 2.0.0
 
 ### Major Changes